JS-5542: settings pages update
